Move the bulk of the new job page into a new job editor component
